div4/div4b/Webwidgets trunk
Fix minor WW bug. I hope.

This patch makes sure that the node property of the ChildNodes class is set before calling the super constructor. This is needed since the super constructor calls a validation method which relies on the node property. I haven't really investigated the code, but hopefully this is a sane fix.