div4/div4b/Webwidgets trunk
Fix another minor WW bug. I hope.

When a row is deleted from a GBOList, the ids of cell elements are not recalulated properly. This causes the cells to be printed with the wrong id, wreaking havoc with all JavaScripts and CSS stylings. This patch is a naive and possibly wrong fix for this issue, it harcodes the attribute drawing function to always recalculate the id. Egil should look at this and possibly make a better patch.